requirements.txt
pandas
requests
openai
langchain
langchain-openai
langgraph
chromadb
tiktoken
python-dotenv
